Move adapters out of fragment
